+++ src/modules/graphql-actions/graphql-action.service.js
@@ -1,13 +1,14 @@
 import axios from 'axios';
+
 import { getDataForAuth } from '../../common/utils';
 import { API_URL } from '../../config';
 
-class CompanyService {
+class GraphqlActionService {
   constructor() {
     this.baseUrl = API_URL;
   }
 
-  async getCompany() {
+  async findAll() {
     const dataForAuth = getDataForAuth();
 
     if (!dataForAuth) {
@@ -17,85 +18,10 @@ class CompanyService {
     const { accessToken, companyUuid } = dataForAuth;
 
     const response = await axios({
-      url: `${this.baseUrl}companies/your-company/${companyUuid}`,
+      url: `${this.baseUrl}graphql-actions/${companyUuid}`,
       headers: {
         Authorization: `Bearer ${accessToken}`,
         'company-uuid': companyUuid,
-      },
-    });
-
-    const { data } = response;
-
-    return [data];
-  }
-
-  async createCompany(company = {}) {
-    const { name, serviceAccount, firebaseConfig, countryCode } = company;
-
-    const body = {
-      name,
-      serviceAccount,
-      firebaseConfig,
-      countryCode
-    };
-
-    const response = await axios({
-      url: `${this.baseUrl}companies`,
-      method: 'post',
-      data: {
-        ...body,
-      },
-    });
-
-    const { data } = response;
-
-    return {
-      ...data,
-      message: `company created, the UUID is ${data.uuid} please save that value.`
-    };
-  }
-
-  async updateCompany(company) {
-    const dataForAuth = getDataForAuth();
-
-    if (!dataForAuth) {
-      throw new Error('can not get data for auth.');
-    }
-
-    const { accessToken, companyUuid } = dataForAuth;
-
-    const {
-      id,
-      name,
-      serviceAccount,
-      firebaseConfig,
-      countryCode,
-      confirmationEmailConfig,
-      forgottenPasswordConfig,
-      logoUrl
-    } = company;
-
-    const body = {
-      name,
-      serviceAccount,
-      firebaseConfig,
-      countryCode,
-      confirmationEmailConfig,
-      forgottenPasswordConfig,
-      logoUrl
-    };
-
-    console.log(body, 'body');
-
-    const response = await axios({
-      url: `${this.baseUrl}companies/${companyUuid}`,
-      method: 'patch',
-      headers: {
-        Authorization: `Bearer ${accessToken}`,
-        'company-uuid': companyUuid,
-      },
-      data: {
-        ...body,
       },
     });
 
@@ -104,7 +30,8 @@ class CompanyService {
     return data;
   }
 
-  async removeCompany(id) {
+  async create(item = {}) {
+    console.log('item', JSON.stringify(item));
     const dataForAuth = getDataForAuth();
 
     if (!dataForAuth) {
@@ -113,13 +40,86 @@ class CompanyService {
 
     const { accessToken, companyUuid } = dataForAuth;
 
+    const { name, isQuery, isMutation, projectId } = item;
+
+    const body = {
+      name,
+      companyUuid,
+      isQuery,
+      isMutation,
+      projectId
+    };
+
     const response = await axios({
-      url: `${this.baseUrl}companies/${companyUuid}`,
-      method: 'delete',
+      url: `${this.baseUrl}graphql-actions`,
+      method: 'post',
       headers: {
         Authorization: `Bearer ${accessToken}`,
         'company-uuid': companyUuid,
       },
+      data: {
+        ...body
+      }
+    });
+
+    const { data } = response;
+
+    return data;
+  }
+
+  async update(item = {}) {
+    const dataForAuth = getDataForAuth();
+
+    if (!dataForAuth) {
+      throw new Error('can not get data for auth.');
+    }
+
+    const { accessToken, companyUuid } = dataForAuth;
+
+    const { id, name, isQuery, isMutation, projectId } = item;
+
+    const body = {
+      name,
+      isQuery,
+      isMutation,
+      projectId
+    };
+
+    const response = await axios({
+      url: `${this.baseUrl}graphql-actions/${companyUuid}/${id}`,
+      method: 'patch',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+        'company-uuid': companyUuid,
+      },
+      data: {
+        ...body
+      }
+    });
+
+    const { data } = response;
+
+    return data;
+  }
+
+  async remove(item = {}) {
+    const dataForAuth = getDataForAuth();
+
+    if (!dataForAuth) {
+      throw new Error('can not get data for auth.');
+    }
+
+    const { accessToken, companyUuid } = dataForAuth;
+
+    const { id } = item;
+
+    const response = await axios({
+      url: `${this.baseUrl}graphql-actions/${companyUuid}/${id}`,
+      method: 'delete',
+      headers: {
+        Authorization: `Bearer ${accessToken}`,
+        'company-uuid': companyUuid,
+      }
     });
 
     const { data } = response;
@@ -128,4 +128,4 @@ class CompanyService {
   }
 }
 
-export const companiesService = new CompanyService();
+export const graphqlActionService = new GraphqlActionService();
